Fix crash in python3 when trying to reload config
reload method in python2 is no longer available in python3. The call in
gui.py defaultValues(self) uses it to restore the config parameters.
Practically, from the parameter settings, when pushing "Default", easyleed
crashes when used with Python3. The proposed patch fixes it by using the
new implementation of reload through imp.reload. This has also being backported
to python2, so it should be safe for both.