Use parser for storing software config parameters
Currently configuration parameters are stored in a module (defaultconfig.py). This makes parameter settings very hard to change dynamically, and requires "recompilation" of the package for any change.The proposed patch uses a parser to produce/read/save a INI file with the parameters. If it's not present already, the INI files is first created in $HOME. In the settings panel, loading and saving parameters are now done by saving copies of the INI files. Apply settings, updates the INI files as well.

This is a proposed patch to fix #9, and provide a much more robust and flexible way with deailing with defaults parameters.