Set parameters after editing without saving. Apply button changed to Save as default
The apply button is confusing. Editing a parameter should be internally tracked (i.e. variables set,without writing on the INI) after editing, without requiring pushing "Apply". This patch enacts this. The Apply button is changed to "Save as Default", where the changes are saved in the INI file. Pusing "Restore" reverts to the factory values.