Fix and streamline saving intensities as csv
Currently intensities are supposedly saved as csv. In reality, their format is not properly csv as there is no comma separating the values. This patch fixes that. Furthermore, rather than having two separate files for the intensity and the average (the latter with an non-standard extension), this patch creates a intensity.csv or a intensity_with-average.csv depending if the checkbox is selected for the average in plot. As the name suggests, it will include or not the average. Finally, this patch makes the header a lot more useful, by adding the correct number of peaks, and the average if selected.