Use pandas dataframe for saving intensities and spot locations
Currently intensities are internally saved as numpy arrays, making it difficult to add other variables (such as center coordinates, spot location and size). This patch changes this by using pandas dataframes. The resulting code is cleaner, more transparent and allows for traking all of the above. In the future it will be used to save independently also spot locations and center coordinates replacing the current binary pkl files.