Replace spot and center locations save/load methods with csv through dataframe
Currently, saving and loading from file spot and center locations are done with pickle using binary files. These are inconveniently useful only to reload them up in Easyleed. This patch uses the internal dataframe to export the locations into a CSV file. This allows for full modification of the file, and more transparent handling of the parameters.