Fix crash when saving center before first acquisition
When a center is set, after opening some images and before the acquisition, easyleed crashes when trying to save the center. This is because the worker has not be set yet. This patch fixes this by checking if the worker exists, and if not, it saves the center from the current image.